reorganize deployments gql
